renamded types to constants
